pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = "A modern Python project template with uv, Hydra, Docker, and dev container support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "lxml>=5.3.2",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pytest>=8.4.2",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
